+++ cmd/apps/substreams_common.go
@@ -1,13 +1,9 @@
 package apps
 
 import (
-	"fmt"
 	"sync"
 
 	"github.com/spf13/cobra"
-	firecore "github.com/streamingfast/firehose-core"
-	"github.com/streamingfast/substreams/pipeline"
-	"github.com/streamingfast/substreams/wasm"
 )
 
 var registerSSOnce sync.Once
@@ -18,23 +14,4 @@ func registerCommonSubstreamsFlags(cmd *cobra.Command) {
 		cmd.Flags().String("substreams-state-store-url", "{sf-data-dir}/localdata", "where substreams state data are stored")
 		cmd.Flags().String("substreams-state-store-default-tag", "", "If non-empty, will be appended to {substreams-state-store-url} (ex: 'v1'). Can be overriden per-request with 'X-Sf-Substreams-Cache-Tag' header")
 	})
-}
-
-func getSubstreamsExtensions[B firecore.Block](chain *firecore.Chain[B]) ([]wasm.WASMExtensioner, []pipeline.PipelineOptioner, error) {
-	var wasmExtensions []wasm.WASMExtensioner
-	var pipelineOptions []pipeline.PipelineOptioner
-
-	if chain.RegisterSubstreamsExtensions != nil {
-		extensions, err := chain.RegisterSubstreamsExtensions(chain)
-		if err != nil {
-			return nil, nil, fmt.Errorf("register substreams extensions failed: %w", err)
-		}
-
-		for _, extension := range extensions {
-			wasmExtensions = append(wasmExtensions, extension.WASMExtensioner)
-			pipelineOptions = append(pipelineOptions, extension.PipelineOptioner)
-		}
-	}
-
-	return wasmExtensions, pipelineOptions, nil
 }