+++ cmd/apps/substreams_tier1.go
@@ -28,6 +28,7 @@ import (
 	"github.com/streamingfast/firehose-core/launcher"
 	"github.com/streamingfast/logging"
 	app "github.com/streamingfast/substreams/app"
+	"github.com/streamingfast/substreams/wasm"
 	"go.uber.org/zap"
 )
 
@@ -97,13 +98,21 @@ func RegisterSubstreamsTier1App[B firecore.Block](chain *firecore.Chain[B], root
 				}
 			}
 
-			wasmExtensions, pipelineOptioner, err := getSubstreamsExtensions(chain)
-			if err != nil {
-				return nil, fmt.Errorf("substreams extensions: %w", err)
+			var wasmExtensions wasm.WASMExtensioner
+			if chain.RegisterSubstreamsExtensions != nil {
+				exts, err := chain.RegisterSubstreamsExtensions()
+				if err != nil {
+					return nil, fmt.Errorf("substreams extensions: %w", err)
+				}
+				wasmExtensions = exts
 			}
+
+			meteringConfig := viper.GetString("common-metering-plugin")
 
 			return app.NewTier1(appLogger,
 				&app.Tier1Config{
+					MeteringConfig: meteringConfig,
+
 					MergedBlocksStoreURL: mergedBlocksStoreURL,
 					OneBlocksStoreURL:    oneBlocksStoreURL,
 					ForkedBlocksStoreURL: forkedBlocksStoreURL,
@@ -117,8 +126,7 @@ func RegisterSubstreamsTier1App[B firecore.Block](chain *firecore.Chain[B], root
 					SubrequestsInsecure:  subrequestsInsecure,
 					SubrequestsPlaintext: subrequestsPlaintext,
 
-					WASMExtensions:  wasmExtensions,
-					PipelineOptions: pipelineOptioner,
+					WASMExtensions: wasmExtensions,
 
 					Tracing: tracing,