+++ CHANGELOG.md
@@ -8,6 +8,10 @@ Operators, you should copy/paste content of this content straight to your projec
 
 If you were at `firehose-core` version `1.0.0` and are bumping to `1.1.0`, you should copy the content between those 2 version to your own repository, replacing placeholder value `fire{chain}` with your chain's own binary.
 
+## v1.3.1
+
+* fix panic on substreams-tier2 service
+
 ## v1.3.0
 
 ### Substreams